fix: pull request adjustments
